blobcache: general single inflight io support
support both digested & indexed
