+++ storage/src/cache/chunkmap/indexed.rs
@@ -2,24 +2,17 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use std::collections::HashMap;
 use std::fs::OpenOptions;
 use std::io::Result;
 use std::os::unix::io::AsRawFd;
-use std::sync::{
-    atomic::{AtomicU8, Ordering},
-    Arc, Condvar, Mutex, WaitTimeoutResult,
-};
-use std::time::Duration;
+use std::sync::atomic::{AtomicU8, Ordering};
 
 use nydus_utils::div_round_up;
 
 use super::ChunkMap;
+use crate::cache::chunkmap::{ChunkIndexGetter, NoWaitSupport};
 use crate::device::RafsChunkInfo;
 use crate::utils::readahead;
-use crate::{StorageError, StorageResult};
-
-use crate::cache::blobcache::SINGLE_INFLIGHT_WAIT_TIMEOUT;
 
 /// The magic number of blob chunk_map file, it's ASCII hex of string "BMAP".
 const MAGIC: u32 = 0x424D_4150;
@@ -51,57 +44,13 @@ pub struct IndexedChunkMap {
     chunk_count: u32,
     size: usize,
     base: *const u8,
-    inflight_tracer: Arc<Mutex<HashMap<u32, Arc<ChunkSlot>>>>,
-}
-
-#[derive(PartialEq)]
-enum Status {
-    Inflight,
-    Complete,
-}
-
-struct ChunkSlot {
-    on_trip: Mutex<Status>,
-    condvar: Condvar,
-}
-
-impl ChunkSlot {
-    fn new() -> Self {
-        ChunkSlot {
-            on_trip: Mutex::new(Status::Inflight),
-            condvar: Condvar::new(),
-        }
-    }
-
-    fn notify(&self) {
-        self.condvar.notify_all();
-    }
-
-    fn done(&self) {
-        // Not expect poisoned lock here
-        *self.on_trip.lock().unwrap() = Status::Complete;
-        self.notify();
-    }
-
-    fn wait_for_inflight(&self, timeout: Duration) -> StorageResult<()> {
-        let mut inflight = self.on_trip.lock().unwrap();
-        let mut tor: WaitTimeoutResult;
-        while *inflight == Status::Inflight {
-            // Do not expect poisoned lock, so unwrap here.
-            let r = self.condvar.wait_timeout(inflight, timeout).unwrap();
-            inflight = r.0;
-            tor = r.1;
-            if tor.timed_out() {
-                return Err(StorageError::Timeout);
-            }
-        }
-
-        Ok(())
-    }
 }
 
 unsafe impl Send for IndexedChunkMap {}
+
 unsafe impl Sync for IndexedChunkMap {}
+
+impl NoWaitSupport for IndexedChunkMap {}
 
 impl IndexedChunkMap {
     pub fn new(blob_path: &str, chunk_count: u32) -> Result<Self> {
@@ -170,7 +119,6 @@ impl IndexedChunkMap {
             chunk_count,
             size: expected_size as usize,
             base: base as *const u8,
-            inflight_tracer: Arc::new(Mutex::new(HashMap::new())),
         })
     }
 
@@ -223,44 +171,19 @@ impl Drop for IndexedChunkMap {
     }
 }
 
+impl ChunkIndexGetter for IndexedChunkMap {
+    type Index = u32;
+
+    fn get_index(chunk: &dyn RafsChunkInfo) -> Self::Index {
+        chunk.blob_index()
+    }
+}
+
 impl ChunkMap for IndexedChunkMap {
-    fn has_ready(&self, chunk: &dyn RafsChunkInfo, wait: bool) -> Result<bool> {
+    fn has_ready(&self, chunk: &dyn RafsChunkInfo, _wait: bool) -> Result<bool> {
         let index = chunk.index();
         let _ = self.validate_index(index)?;
         let (ready, _) = self.is_chunk_ready(index);
-        if !ready {
-            let mut guard = self.inflight_tracer.lock().unwrap();
-            trace!(
-                "chunk index {}, tracer scale {}",
-                chunk.index(),
-                guard.len()
-            );
-            if let Some(i) = guard.get(&index).cloned() {
-                if wait {
-                    drop(guard);
-                    match i.wait_for_inflight(Duration::from_millis(SINGLE_INFLIGHT_WAIT_TIMEOUT)) {
-                        Err(StorageError::Timeout) => {
-                            // Notice that lock of tracer is already dropped.
-                            let mut t = self.inflight_tracer.lock().unwrap();
-                            t.remove(&index);
-                            i.notify();
-                            warn!("Waiting for another backend IO expires. chunk index {}, tracer scale {}", chunk.index(), t.len());
-                            return Ok(self.is_chunk_ready(chunk.index()).0);
-                        }
-                        _ => {
-                            return Ok(self.is_chunk_ready(chunk.index()).0);
-                        }
-                    };
-                }
-            } else {
-                // Double check to close the window where prior slot was just
-                // removed after backend IO returned.
-                if self.is_chunk_ready(index).0 {
-                    return Ok(true);
-                }
-                guard.insert(index, Arc::new(ChunkSlot::new()));
-            }
-        }
         Ok(ready)
     }
 
@@ -279,127 +202,6 @@ impl ChunkMap for IndexedChunkMap {
                 break;
             }
         }
-
-        self.finish(chunk);
-
         Ok(())
-    }
-
-    fn finish(&self, chunk: &dyn RafsChunkInfo) {
-        let index = chunk.index();
-        let mut guard = self.inflight_tracer.lock().unwrap();
-        if let Some(i) = guard.remove(&index) {
-            i.done();
-        }
-    }
-}
-
-#[cfg(test)]
-mod this_test {
-    use std::sync::Arc;
-    use std::thread;
-    use std::time::Duration;
-
-    use vmm_sys_util::tempfile::TempFile;
-
-    use super::IndexedChunkMap;
-    use crate::cache::blobcache::blob_cache_tests::MockChunkInfo;
-    use crate::cache::chunkmap::ChunkMap;
-    use crate::device::RafsChunkInfo;
-
-    #[test]
-    fn test_inflight_tracer_race() {
-        let tmp_file = TempFile::new().unwrap();
-        let map = Arc::new(IndexedChunkMap::new(tmp_file.as_path().to_str().unwrap(), 10).unwrap());
-
-        let chunk_4: Arc<dyn RafsChunkInfo> = Arc::new({
-            let mut c = MockChunkInfo::new();
-            c.index = 4;
-            c
-        });
-
-        map.as_ref().has_ready(chunk_4.as_ref(), false).unwrap();
-        let map_cloned = map.clone();
-
-        assert_eq!(map.inflight_tracer.lock().unwrap().len(), 1);
-
-        let chunk_4_cloned = chunk_4.clone();
-        let t1 = thread::Builder::new()
-            .spawn(move || {
-                for _ in 0..4 {
-                    let ready = map_cloned.has_ready(chunk_4_cloned.as_ref(), true).unwrap();
-                    assert_eq!(ready, true);
-                }
-            })
-            .unwrap();
-
-        let map_cloned_2 = map.clone();
-        let chunk_4_cloned_2 = chunk_4.clone();
-        let t2 = thread::Builder::new()
-            .spawn(move || {
-                for _ in 0..2 {
-                    let ready = map_cloned_2
-                        .has_ready(chunk_4_cloned_2.as_ref(), true)
-                        .unwrap();
-                    assert_eq!(ready, true);
-                }
-            })
-            .unwrap();
-
-        thread::sleep(Duration::from_secs(1));
-
-        map.set_ready(chunk_4.as_ref()).unwrap();
-
-        // Fuzz
-        map.set_ready(chunk_4.as_ref()).unwrap();
-        map.set_ready(chunk_4.as_ref()).unwrap();
-
-        assert_eq!(map.inflight_tracer.lock().unwrap().len(), 0);
-
-        t1.join().unwrap();
-        t2.join().unwrap();
-    }
-
-    #[test]
-    /// Case description:
-    ///     Never invoke `set_ready` method, thus to let each caller of `has_ready` reach
-    ///     a point of timeout.
-    /// Expect:
-    ///     The chunk of index 4 is never marked as ready/downloaded.
-    ///     Each caller of `has_ready` can escape from where it is blocked.
-    ///     After timeout, no slot is left in inflight tracer.
-    fn test_inflight_tracer_timeout() {
-        let tmp_file = TempFile::new().unwrap();
-        let map = Arc::new(IndexedChunkMap::new(tmp_file.as_path().to_str().unwrap(), 10).unwrap());
-
-        let chunk_4: Arc<dyn RafsChunkInfo> = Arc::new({
-            let mut c = MockChunkInfo::new();
-            c.index = 4;
-            c
-        });
-
-        map.as_ref().has_ready(chunk_4.as_ref(), false).unwrap();
-        let map_cloned = map.clone();
-
-        assert_eq!(map.inflight_tracer.lock().unwrap().len(), 1);
-
-        let chunk_4_cloned = chunk_4.clone();
-        let t1 = thread::Builder::new()
-            .spawn(move || {
-                for _ in 0..4 {
-                    map_cloned.has_ready(chunk_4_cloned.as_ref(), true).unwrap();
-                }
-            })
-            .unwrap();
-
-        t1.join().unwrap();
-
-        assert_eq!(map.inflight_tracer.lock().unwrap().len(), 1);
-
-        let ready = map.as_ref().has_ready(chunk_4.as_ref(), false).unwrap();
-        assert_eq!(ready, false);
-
-        map.finish(chunk_4.as_ref());
-        assert_eq!(map.inflight_tracer.lock().unwrap().len(), 0);
     }
 }
